+++ pyo3-object_store/src/aws/mod.rs
@@ -1,4 +1,6 @@
 mod credentials;
+mod remote_signed;
 mod store;
 
+pub use remote_signed::PyRemoteSignedS3Store;
 pub use store::PyS3Store;